% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_transform.R
\name{pw_transform}
\alias{pw_transform}
\title{Prais-Winsten Transformation}
\usage{
pw_transform(data, rho, intercept = TRUE, groups = NULL)
}
\arguments{
\item{data}{a named matrix containing the data that should be transformed.}

\item{rho}{the AR(1) serial correlation coefficient \eqn{\rho}.}

\item{intercept}{logal. If TRUE (default) a column of intercept values is added.}

\item{groups}{a list containing a vector of positions of each
ID variable as its elements. Only required for panel data.}
}
\value{
A named matrix.
}
\description{
Performs the Prais-Winsten transformation for a given data set.
}
