% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.prais.R, R/print.summary.prais.R
\name{summary.prais}
\alias{summary.prais}
\alias{print.summary.prais}
\title{Summarising the Prais-Winsten Estimator}
\usage{
\method{summary}{prais}(object, ...)

\method{print}{summary.prais}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"prais"}, usually, a result of a call to
\code{\link{prais_winsten}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"summary.prais"}, usually, a result of a call to
\code{\link{summary.prais}}.}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If \code{TRUE}, 'significance stars' are printed
for each coefficient.}
}
\value{
\code{summary.prais} returns a list of class \code{"summary.prais"},
which contains the following components:
\item{call}{the matched call.}
\item{residuals}{the residuals of the Prais-Winsten transformed model, i.e. the
residuals the reported standard errors are based on. The residuals on the scale
of the original data are available as \code{residuals} of the estimated object.}
\item{coefficients}{a named vector of coefficients.}
\item{rho}{the values of the AR(1) coefficient \eqn{\rho} from all iterations.}
\item{sigma}{the square root of the estimated variance of the random error.}
\item{df}{degrees of freedom, a 3-vector \emph{(p, n-p, p*)},
the first being the number of non-aliased coefficients, the last being
the total number of coefficients.}
\item{r.squared}{R^2, the 'fraction of variance explained by the model',
\deqn{R^2 = 1 - \frac{\sum {(y_i - \hat{y}_i)^2}}{\sum {(y_i - \overline{y})^2}},}
where \eqn{\overline{y}} is the mean of \eqn{y_i} for \eqn{y_i = 1, ..., N}
if there is an intercept and zero otherwise.}
\item{adj.r.squared}{the above \emph{R^2} statistic \emph{'adjusted'}, penalising
for higher \emph{p}.}
\item{fstatistic}{(for models including non-intercept terms) a 3-vector with the
value of the F-statistic with its numerator and denominator degrees of freedom.}
\item{cov.unscaled}{a \eqn{p \times p} matrix of (unscaled) covariances of the
\emph{coef[j], j=1, ..., p}.}
\item{dw}{a named 2-vector with the Durbin-Watson statistic of the original
linear model and the Prais-Winsten estimator.}
\item{terms}{the terms object used.}
}
\description{
Summary method for class \code{"prais"}.
}
\examples{
# Generate an artificial sample
set.seed(1234567)
n <- 100
x <- sample(20:40, n, replace = TRUE)
rho <- .5

# AR(1) errors
u <- rnorm(n, 0, 5)
for (i in 2:n) {
  u[i] <- u[i] + rho * u[i - 1]
}
pw_sample <- data.frame("x" = x, "y" = 10 + 1.5 * x + u, "time" = 1:n)

# Estimate
pw <- prais_winsten(y ~ x, data = pw_sample, index = "time")

# The coefficient table, the AR(1) coefficient and the Durbin-Watson
# statistics of the original and of the transformed model
summary(pw)

# The standard errors are those of 'vcov'
all.equal(sqrt(diag(vcov(pw))),
          summary(pw)$coefficients[, "Std. Error"])

}
