% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.prais}
\alias{glance.prais}
\title{Glance}
\usage{
\method{glance}{prais}(x, ...)
}
\arguments{
\item{x}{A `prais` object returned from [prais::prais_winsten()].}

\item{...}{Additional arguments. Not used. Needed to match generic
  signature only. **Cautionary note:** Misspelled arguments will be
  absorbed in `...`, where they will be ignored. If the misspelled
  argument has a default value, the default value will be used.
  For example, if you pass `conf.lvel = 0.9`, all computation will
  proceed using `conf.level = 0.95`.

# load libraries for models and data
library(prais)
library(wooldridge)
data("barium")

pw <- prais_winsten(lchnimp ~ lchempi + lgas + lrtwex + befile6 + affile6 + afdec6,
data = barium, index = "t")

tidy(pw)

glance(pw)}
}
\description{
Glance
}
\seealso{
[glance()], [prais::prais_winsten()]
}
\concept{prais_tidiers}
