% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prais-package.R
\docType{package}
\name{prais-package}
\alias{prais}
\alias{prais-package}
\title{prais: Prais-Winsten Estimator for AR(1) Serial Correlation}
\description{
The Prais-Winsten estimator (Prais & Winsten, 1954) takes into account AR(1) serial correlation of the errors in a linear regression model. The procedure recursively estimates the coefficients and the error autocorrelation of the specified model until sufficient convergence of the AR(1) coefficient is attained.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franzmohr/prais}
  \item Report bugs at \url{https://github.com/franzmohr/prais/issues}
}

}
\author{
\strong{Maintainer}: Franz X. Mohr \email{franz.x.mohr@outlook.com} (\href{https://orcid.org/0009-0003-8890-7781}{ORCID})

Authors:
\itemize{
  \item Franz X. Mohr \email{franz.x.mohr@outlook.com} (\href{https://orcid.org/0009-0003-8890-7781}{ORCID})
}

Other contributors:
\itemize{
  \item Michael Škvrňák (\href{https://orcid.org/0000-0003-4647-4095}{ORCID}) [contributor]
}

}
\keyword{internal}
