% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prais.winsten.R
\name{prais.winsten}
\alias{prais.winsten}
\title{Prais-Winsten Estimator for AR(1) Serial Correlation}
\usage{
prais.winsten(formula, max_iter = 50L, tol = 1e-06, twostep = FALSE, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{max_iter}{integer specifying the maximum number of iterations. The default is 50.}

\item{tol}{numeric specifying the maximum absolute difference between the estimator of rho in the current and the previous iteration that has to be attained to reach convergence. The default is 1e-6.}

\item{twostep}{logical. If \code{TRUE}, the estimation will stop after the first iteration.}

\item{...}{additional arguments to be passed to \code{lm}.}
}
\description{
The Prais-Winsten estimator takes into account AR(1) serial correlation of the errors
in a linear regression model. The procedure recursively estimates the coefficients
and the error autocorrelation of the specified model until sufficient convergence of
the AR(1) coefficient is attained. All estimates are obtained by OLS.
}
\references{
Prais, S. J. and Winsten, C. B. (1954): Trend Estimators and Serial Correlation. Cowles Commission Discussion Paper, 383 (Chicago).

Wooldridge, J. M. (2013): Introductory Econometrics. A Modern Approach. 5th ed. Mason, OH: South-Western Cengage Learning Cengage.
}
