% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prais.winsten.R
\name{prais.winsten}
\alias{prais.winsten}
\title{Prais-Winsten Estimator for AR(1) Serial Correlation}
\usage{
prais.winsten(formula, max_iter = 50, tol = 1e-06, twostep = FALSE, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{max_iter}{integer specifying the maximum number of iterations.}

\item{tol}{double the maximum (positive) value of the absolute difference between the estimator of rho in the current and the previous iteration that has to be attained to reach convergence and to stop the estimation. If not specified the value 1e-06 is used.}

\item{twostep}{logical specifying if estimation should stop after the first iteration.}

\item{...}{further arguments passed on to \code{lm}.}
}
\description{
The Prais-Winsten estimator takes into account serial correlation of type AR(1) in a linear model.
The procedure recursively estimates the beta coefficients and the error autocorrelation
of the specified model until sufficient convergence of rho, i.e. the AR(1) coefficient, is attained. All estimates are obtained by OLS.
}
\references{
Prais, S. J. and Winsten, C. B. (1954): Trend Estimators and Serial Correlation. Cowles Commission Discussion Paper, 383 (Chicago).

Wooldridge, J. M. (2013): Introductory Econometrics. A Modern Approach. 5th ed. Mason, OH: South-Western Cengage Learning Cengage.
}
