% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.prais.R
\name{predict.prais}
\alias{predict.prais}
\title{Predict Method for Objects of Class prais}
\usage{
\method{predict}{prais}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"prais"}, usually, a result of a call to
\code{\link{prais_winsten}}.}

\item{newdata}{an optional data frame in which to look for variables with which to
predict. It must contain the variables that appear in \code{formula}, which do not
have to be transformed beforehand. If omitted, the fitted values are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector of predictions.
}
\description{
Predicted values based on Prais-Winsten object.
}
\details{
The predictions are the conditional mean of the model, i.e. the product
of the regressors and the coefficients. The AR(1) structure of the errors is not
used, so the result does not contain the forecast of the serially correlated part
of the error term.
}
\examples{
# Generate an artificial sample
set.seed(1234567)
n <- 100
x <- sample(20:40, n, replace = TRUE)
rho <- .5

# AR(1) errors
u <- rnorm(n, 0, 5)
for (i in 2:n) {
  u[i] <- u[i] + rho * u[i - 1]
}
pw_sample <- data.frame("x" = x, "y" = 10 + 1.5 * x + u, "time" = 1:n)

# Estimate
pw <- prais_winsten(y ~ x, data = pw_sample, index = "time")

# Predict
fcst <- predict(pw)

}
\references{
Prais, S. J. and Winsten, C. B. (1954): Trend Estimators and Serial Correlation. Cowles Commission Discussion Paper, 383 (Chicago).
}
