% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_transform.R
\name{pw_transform}
\alias{pw_transform}
\title{Prais-Winsten Transformation}
\usage{
pw_transform(data, rho, intercept = TRUE)
}
\arguments{
\item{data}{a named matrix containing the data, which should be transformed.}

\item{rho}{a numeric of the AR(1) serial correlation coefficient \eqn{\rho}.}

\item{intercept}{logal. If TRUE (default) a column of intercept values is added.}
}
\value{
a named matrix.
}
\description{
Performs the Prais-Winsten transformation for a given data set.
}
