% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.prais}
\alias{tidy.prais}
\title{Tidy}
\usage{
\method{tidy}{prais}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{an object of class `prais`, usually, the result of a call to \code{\link{prais_winsten}}.}

\item{conf.int}{logical indicating whether or not to include
a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{the confidence level to use for the confidence
interval if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{additional arguments. Not used. Needed to match generic
  signature only. **Cautionary note:** Misspelled arguments will be
  absorbed in \code{...}, where they will be ignored. If the misspelled
  argument has a default value, the default value will be used.
  For example, if you pass \code{conf.lvel = 0.9}, all computation will
  proceed using \code{conf.level = 0.95}.

# load libraries for models and data
library(prais)
library(wooldridge)
data("barium")

pw <- prais_winsten(lchnimp ~ lchempi + lgas + lrtwex + befile6 + affile6 + afdec6,
data = barium, index = "t")

tidy(pw)

glance(pw)}
}
\description{
Tidy
}
