% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.prais.R
\name{vcov.prais}
\alias{vcov.prais}
\title{Covariance Matrix of the Coefficients of the Prais-Winsten Estimator}
\usage{
\method{vcov}{prais}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"prais"}, usually, a result of a call to
\code{\link{prais_winsten}}.}

\item{...}{not used.}
}
\value{
An object of class "matrix" containing the estimate of the covariance
matrix of the coefficients.
}
\description{
Returns the estimated covariance matrix of the coefficients of a model of
class \code{"prais"}.
}
\details{
The covariance matrix is obtained from the Prais-Winsten transformed
model. It is the matrix the standard errors of \code{\link{summary.prais}} are
based on and is identical to \code{vcovHC(object, type = "const")}. Robust
alternatives are available in \code{\link{vcovHC.prais}} and
\code{\link{vcovPC.prais}}.

Coefficients of linearly dependent variables are \code{NA} and are omitted, as
in the covariance matrix of an object of class \code{"lm"}.
}
\examples{
# Generate an artificial sample
set.seed(1234567)
n <- 100
x <- sample(20:40, n, replace = TRUE)
rho <- .5

# AR(1) errors
u <- rnorm(n, 0, 5)
for (i in 2:n) {
  u[i] <- u[i] + rho * u[i - 1]
}
pw_sample <- data.frame("x" = x, "y" = 10 + 1.5 * x + u, "time" = 1:n)

# Estimate
pw <- prais_winsten(y ~ x, data = pw_sample, index = "time")

# Covariance matrix and confidence intervals
vcov(pw)
confint(pw)

}
\seealso{
\code{\link{vcovHC.prais}}, \code{\link{vcovPC.prais}},
\code{\link{confint.prais}}
}
